feat: catch SIGPIPE via no-op handler (logs to debug)
SIGPIPE is sent to the process when libvirt attempts to write to
a dead socket.

We catch SIGPIPE here, because we want to maintain control over
the application runtime when this occurs -- it simply signals to
us that an operation was attempted over a dead socket.

At this point, libvirt is the only thing sending SIGPIPE.

Signed-off-by: Kevin Morris <kevr@0cost.org>