feat: add virt::connection_pool, a persistent connection manager
This allows us to reuse the same object for connections bound
for a particular user, immensely speeding up queries after the
connection is created.

With SIGPIPE ignored, we can utilize libvirt function binds for
connection closure which toggles connection::closed() off.

In the case where closed() is false, which can be caused by
the libvirt socket being closed (killing the ssh session, etc),
we attempt to reconnect five times.

Signed-off-by: Kevin Morris <kevr@0cost.org>